requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
huggingface-hub==0.19.4
# Note: py-feat has complex dependencies, installing core components separately
facenet-pytorch==2.5.3
opencv-python==4.8.1.78
# Vector database and data processing
chromadb==0.4.22
pandas==2.0.3
